backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-chess==1.999
httpx==0.25.1
python-multipart==0.0.6
